+++ packages/tasit-contracts/test/MANAToken.js
@@ -1,7 +1,8 @@
 // Note:
 // Should we use ethers instead web3?
 contract("MANAToken", function(accounts) {
-  const abi = require("../build/contracts/MANAToken.json").abi;
+  const abi = require("../decentraland/mana/build/contracts/MANAToken.json")
+    .abi;
   const address = "0xb32939da0c44bf255b9810421a55be095f9bb3f4";
 
   it("should get the MANAToken name", async function() {
